#363 get last package script

+++ packages/var-rpm.sh
@@ -9,7 +9,12 @@ fi
 chkconfig --add $bin_name && chkconfig $bin_name on
 "
 
-rpm_postun='if [ `find /opt/echoes-alert/probe -type f | wc -l` -eq 0 ]
+rpm_postun='cd ${PACKAGE_TOP_DIR}
+if [ -d sbin ]
+then
+    rm -rf sbin
+fi
+if [ `find /opt/echoes-alert/probe -type f | wc -l` -eq 0 ]
 then
     rm -rf /opt/echoes-alert/probe
     if [ `find /opt/echoes-alert -type f | wc -l` -eq 0 ]
@@ -18,8 +23,12 @@ then
     fi
 fi'
 
-pm_preun="service $bin_name stop
+rpm_preun="service $bin_name stop
 cd ${PACKAGE_TOP_DIR}
+if [ -d etc ]
+then
+    rm -rf etc
+fi
 if [ -d lib ]
 then
     rm -rf lib
@@ -31,17 +40,14 @@ fi
 chkconfig $bin_name off
 "
 
-rpm_files="\"/${PACKAGE_TOP_DIR}sbin/ea-probe\"
-\"/${PACKAGE_TOP_DIR}etc/probe.conf\"
-\"//var/log/echoes-alert/probe.log\"
-\"//etc/init.d/ea-probe\"
-\"//etc/sysconfig/ea-probe\"
-\"/${PACKAGE_TOP_DIR}doc/AUTHORS\"
-\"/${PACKAGE_TOP_DIR}doc/README\"
-\"/${PACKAGE_TOP_DIR}doc/COPYRIGHT\""
-
-rpm_dir="\"//var/log/echoes-alert/\"
-\"/${PACKAGE_TOP_DIR}etc/plugins\""
+rpm_files="\"/${PACKAGE_TOP_DIR}/sbin/ea-probe\"
+\"/${PACKAGE_TOP_DIR}/etc/probe.conf\"
+\"/var/log/echoes-alert/probe.log\"
+\"/etc/init.d/ea-probe\"
+\"/etc/sysconfig/ea-probe\"
+\"/${PACKAGE_TOP_DIR}/doc/AUTHORS\"
+\"/${PACKAGE_TOP_DIR}/doc/README\"
+\"/${PACKAGE_TOP_DIR}/doc/COPYRIGHT\""
 
 # Sysconfig
 createVarDir "rpm_sysconfig" "${NBTMPDIR}/etc/sysconfig/" 0755
